moved RunCycle logic into SchedulerProfile
