+++ pkg/epp/scheduling/scheduler.go
@@ -30,7 +30,6 @@ import (
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/framework/plugins/picker"
 	profilepicker "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/framework/plugins/profile-picker"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/types"
-	errutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/error"
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/logging"
 )
 
@@ -94,7 +93,7 @@ type Datastore interface {
 }
 
 // Schedule finds the target pod based on metrics and the requested lora adapter.
-func (s *Scheduler) Schedule(ctx context.Context, req *types.LLMRequest) ([]*types.Result, error) {
+func (s *Scheduler) Schedule(ctx context.Context, req *types.LLMRequest) (map[string]*types.Result, error) {
 	logger := log.FromContext(ctx).WithValues("request", req)
 	loggerDebug := logger.V(logutil.DEBUG)
 
@@ -103,128 +102,38 @@ func (s *Scheduler) Schedule(ctx context.Context, req *types.LLMRequest) ([]*typ
 		metrics.RecordSchedulerE2ELatency(time.Since(scheduleStart))
 	}()
 
-	before := time.Now()
-	profiles := s.profilePicker.Pick(req, s.profiles) // pick profiles to run
-	metrics.RecordSchedulerPluginProcessingLatency(framework.ProfilePickerType, s.profilePicker.Name(), time.Since(before))
-
-	if len(profiles) == 0 {
-		return nil, fmt.Errorf("failed to pick SchedulingProfile for the request - %s", req)
-	}
 	// Snapshot pod metrics from the datastore to:
 	// 1. Reduce concurrent access to the datastore.
 	// 2. Ensure consistent data during the scheduling operation of a request between all scheduling cycles.
 	sCtx := types.NewSchedulingContext(ctx, req, nil, types.ToSchedulerPodMetrics(s.datastore.PodGetAll()))
 	loggerDebug.Info(fmt.Sprintf("Scheduling a request, Metrics: %+v", sCtx.PodsSnapshot))
 
-	result := []*types.Result{}
-	for name, profile := range profiles {
-		// run the selected profiles and collect results (current code runs all profiles)
-		cycleResult, err := s.runSchedulerProfileCycle(sCtx, profile)
-		if err != nil {
-			return nil, fmt.Errorf("failed to run all required scheduling profiles - %w", err)
-		}
-		cycleResult.ProfileName = name
-		result = append(result, cycleResult)
-	}
+	profileExecutionResults := map[string]*types.Result{}
 
-	return result, nil
-}
-
-func (s *Scheduler) runSchedulerProfileCycle(ctx *types.SchedulingContext, profile *framework.SchedulerProfile) (*types.Result, error) {
-	s.runPreCyclePlugins(ctx, profile.PreCyclePlugins())
-
-	pods := s.runFilterPlugins(ctx, profile.Filters())
-	if len(pods) == 0 {
-		return nil, errutil.Error{Code: errutil.Internal, Msg: "no pods available for the given request"}
-	}
-	// if we got here, there is at least one pod to score
-	weightedScorePerPod := s.runScorerPlugins(ctx, pods, profile.Scorers())
-
-	result := s.runPickerPlugin(ctx, weightedScorePerPod, profile.Picker())
-
-	s.runPostCyclePlugins(ctx, result, profile.PostCyclePlugins())
-
-	return result, nil
-}
-
-func (s *Scheduler) runPreCyclePlugins(ctx *types.SchedulingContext, plugins []framework.PreCycle) {
-	for _, plugin := range plugins {
-		ctx.Logger.V(logutil.DEBUG).Info("Running pre-cycle plugin", "plugin", plugin.Name())
+	for { // get the next set of profiles to run iteratively based on the request and the previous execution results
 		before := time.Now()
-		plugin.PreCycle(ctx)
-		metrics.RecordSchedulerPluginProcessingLatency(framework.PreCyclePluginType, plugin.Name(), time.Since(before))
-	}
-}
-
-func (s *Scheduler) runFilterPlugins(ctx *types.SchedulingContext, filters []framework.Filter) []types.Pod {
-	loggerDebug := ctx.Logger.V(logutil.DEBUG)
-	filteredPods := ctx.PodsSnapshot
-	loggerDebug.Info("Before running filter plugins", "pods", filteredPods)
-
-	for _, filter := range filters {
-		loggerDebug.Info("Running filter plugin", "plugin", filter.Name())
-		before := time.Now()
-		filteredPods = filter.Filter(ctx, filteredPods)
-		metrics.RecordSchedulerPluginProcessingLatency(framework.FilterPluginType, filter.Name(), time.Since(before))
-		loggerDebug.Info("Filter plugin result", "plugin", filter.Name(), "pods", filteredPods)
-		if len(filteredPods) == 0 {
+		profiles := s.profilePicker.Pick(req, s.profiles, profileExecutionResults)
+		metrics.RecordSchedulerPluginProcessingLatency(framework.ProfilePickerType, s.profilePicker.Name(), time.Since(before))
+		if len(profiles) == 0 { // profile picker didn't pick any profile to run
 			break
 		}
-	}
-	loggerDebug.Info("After running filter plugins")
 
-	return filteredPods
-}
+		for name, profile := range profiles {
+			// run the selected profiles and collect results (current code runs all profiles)
+			profileExecutionResult, err := profile.RunCycle(sCtx)
+			if err != nil {
+				return nil, fmt.Errorf("failed to run all required scheduling profiles - %w", err)
+			}
 
-func (s *Scheduler) runScorerPlugins(ctx *types.SchedulingContext, pods []types.Pod, scorers []*framework.WeightedScorer) map[types.Pod]float64 {
-	loggerDebug := ctx.Logger.V(logutil.DEBUG)
-	loggerDebug.Info("Before running scorer plugins", "pods", pods)
-
-	weightedScorePerPod := make(map[types.Pod]float64, len(pods))
-	for _, pod := range pods {
-		weightedScorePerPod[pod] = float64(0) // initialize weighted score per pod with 0 value
-	}
-	// Iterate through each scorer in the chain and accumulate the weighted scores.
-	for _, scorer := range scorers {
-		loggerDebug.Info("Running scorer", "scorer", scorer.Name())
-		before := time.Now()
-		scores := scorer.Score(ctx, pods)
-		metrics.RecordSchedulerPluginProcessingLatency(framework.ScorerPluginType, scorer.Name(), time.Since(before))
-		for pod, score := range scores { // weight is relative to the sum of weights
-			weightedScorePerPod[pod] += score * float64(scorer.Weight())
+			profileExecutionResults[name] = profileExecutionResult
 		}
-		loggerDebug.Info("After running scorer", "scorer", scorer.Name())
-	}
-	loggerDebug.Info("After running scorer plugins")
-
-	return weightedScorePerPod
-}
-
-func (s *Scheduler) runPickerPlugin(ctx *types.SchedulingContext, weightedScorePerPod map[types.Pod]float64, picker framework.Picker) *types.Result {
-	loggerDebug := ctx.Logger.V(logutil.DEBUG)
-	scoredPods := make([]*types.ScoredPod, len(weightedScorePerPod))
-	i := 0
-	for pod, score := range weightedScorePerPod {
-		scoredPods[i] = &types.ScoredPod{Pod: pod, Score: score}
-		i++
 	}
 
-	loggerDebug.Info("Before running picker plugin", "pods weighted score", fmt.Sprint(weightedScorePerPod))
-	before := time.Now()
-	result := picker.Pick(ctx, scoredPods)
-	metrics.RecordSchedulerPluginProcessingLatency(framework.PickerPluginType, picker.Name(), time.Since(before))
-	loggerDebug.Info("After running picker plugin", "result", result)
-
-	return result
-}
-
-func (s *Scheduler) runPostCyclePlugins(ctx *types.SchedulingContext, res *types.Result, plugins []framework.PostCycle) {
-	for _, plugin := range plugins {
-		ctx.Logger.V(logutil.DEBUG).Info("Running post-cycle plugin", "plugin", plugin.Name())
-		before := time.Now()
-		plugin.PostCycle(ctx, res)
-		metrics.RecordSchedulerPluginProcessingLatency(framework.PostCyclePluginType, plugin.Name(), time.Since(before))
+	if len(profileExecutionResults) == 0 {
+		return nil, fmt.Errorf("failed to run any SchedulingProfile for the request - %s", req)
 	}
+
+	return profileExecutionResults, nil
 }
 
 // OnResponse is invoked during the processing of a response from an inference pod. It will invoke
@@ -245,7 +154,8 @@ func (s *Scheduler) OnResponse(ctx context.Context, resp *types.LLMResponse, tar
 	sCtx := types.NewSchedulingContext(ctx, nil, resp, pods)
 
 	// WORKAROUND until PostResponse is out of Scheduler
-	profiles := s.profilePicker.Pick(nil, s.profiles) // all profiles
+	profileExecutionResults := map[string]*types.Result{}
+	profiles := s.profilePicker.Pick(nil, s.profiles, profileExecutionResults) // all profiles
 	for _, profile := range profiles {
 		s.runPostResponsePlugins(sCtx, targetPod, profile)
 	}
